Rarity Logic Refactored
Rarity logic that was shared between Summaries and Verses has now been pulled out into the new file utils/randomChapter.ts
