+++ app/(tabs)/summaries.tsx
@@ -1,20 +1,13 @@
 import React from 'react';
 import { View, Text, ActivityIndicator } from 'react-native';
 import { ReviewScreenTemplate, ReviewItem } from '@/components/ReviewScreenTemplate';
-import { useBibleBooks, Chapter } from '../../context/BibleBooksContext';
-import type { Rarity } from '../../context/BibleBooksContext';
+import { useBibleBooks } from '../../context/BibleBooksContext';
+import { getWeightedChapters, selectWeightedChapter } from '@/utils/randomChapter';
 
 export default function Summaries() {
   const { bibleBooks } = useBibleBooks();
 
   const enabledBooks = bibleBooks.filter(b => b.Enabled && b.Chapters && b.Chapters.length > 0);
-
-  const rarityWeightMap: Record<Rarity, number> = {
-  common: 1.0,
-  uncommon: 0.5,
-  rare: 0.2,
-  disabled: 0.0,
-};
 
   // If no enabled books with chapters, show loading or info
   if (enabledBooks.length === 0) {
@@ -27,57 +20,11 @@ export default function Summaries() {
   }
 
   const getRandomSummary = async () => {
-    const weightedChapters: {
-      book: string;
-      chapterIndex: number;
-      chapter: Chapter;
-      weight: number;
-    }[] = [];
+    const weightedChapters = getWeightedChapters(enabledBooks);
 
-    for(const book of enabledBooks) {
-      if(!book.Chapters) continue;
+    if(weightedChapters.length === 0) throw new Error('No eligible chapters.');
 
-      for (const chapter of book.Chapters) {
-        const rarity = chapter.rarity ?? 'common';
-        const rarityWeight = rarityWeightMap[rarity];
-
-        if (rarityWeight > 0) {
-          weightedChapters.push({
-            book: book.Book,
-            chapterIndex: chapter.Chapter,
-            chapter,
-            weight: rarityWeight,
-          });
-        }
-      }
-    }
-
-    if (weightedChapters.length === 0) {
-      throw new Error('No eligible chapters with non-zero weight.')
-    }
-
-    const totalWeight = weightedChapters.reduce((sum, ch) => sum + ch.weight, 0);
-    const rand = Math.random() * totalWeight;
-
-    let runningWeight = 0;
-    let selected = weightedChapters[0];
-
-    for (const entry of weightedChapters) {
-      runningWeight += entry.weight;
-      if (rand <= runningWeight) {
-        selected = entry;
-        break;
-      }
-    }
-
-    const { book, chapter, chapterIndex } = selected;
-
-    const randomBook = enabledBooks[Math.floor(Math.random() * enabledBooks.length)];
-    const chapters = randomBook.Chapters;
-
-    if (!chapters || chapters.length === 0) {
-      throw new Error(`No chapters in selected book: ${randomBook.Book}`);
-    }
+    const { book, chapter, chapterIndex } = selectWeightedChapter(weightedChapters);
 
     return {
       book,
