+++ app/(tabs)/verses.tsx
@@ -2,19 +2,12 @@ import React from 'react';
 import { View, Text, ActivityIndicator } from 'react-native';
 import { ReviewScreenTemplate } from '../../components/ReviewScreenTemplate';
 import { useBibleBooks, Chapter } from '../../context/BibleBooksContext';
-import type { Rarity } from '../../context/BibleBooksContext';
+import { getWeightedChapters, selectWeightedChapter } from '@/utils/randomChapter';
 
 export default function Verses() {
     const { bibleBooks } = useBibleBooks();
   
     const enabledBooks = bibleBooks.filter(b => b.Enabled && b.Chapters && b.Chapters.length > 0);
-
-    const rarityWeightMap: Record<Rarity, number> = {
-      common: 1.0,
-      uncommon: 0.5,
-      rare: 0.2,
-      disabled: 0.0,
-    }
   
     // If no enabled books with chapters, show loading or info
     if (enabledBooks.length === 0) {
@@ -26,50 +19,11 @@ export default function Verses() {
       );
     }
   const getRandomVerse = async () => {
-    const weightedChapters: {
-      book: string;
-      chapterIndex: number;
-      chapter: Chapter;
-      weight: number;
-    }[] = [];
+    const weightedChapters = getWeightedChapters(enabledBooks);
 
-    for (const book of enabledBooks) {
-      if (!book.Chapters) continue;
+    if(weightedChapters.length === 0) throw new Error('No eligible chapters.');
 
-      for (const chapter of book.Chapters) {
-        const rarity = chapter.rarity ?? 'common';
-        const rarityWeight = rarityWeightMap[rarity];
-
-        if(rarityWeight > 0) {
-          weightedChapters.push({
-            book: book.Book,
-            chapterIndex: chapter.Chapter,
-            chapter,
-            weight: rarityWeight
-          });
-        }
-      }
-    }
-
-    if (weightedChapters.length === 0) {
-      throw new Error('No eligible chapters with non-zero weight.');
-    }
-
-    const totalWeight = weightedChapters.reduce((sum, ch) => sum + ch.weight, 0);
-    const rand = Math.random() * totalWeight;
-
-    let runningWeight = 0;
-    let selected = weightedChapters[0];
-
-    for (const entry of weightedChapters) {
-      runningWeight += entry.weight;
-      if (rand <= runningWeight) {
-        selected = entry;
-        break;
-      }
-    }
-
-    const { book, chapter, chapterIndex } = selected;
+    const { book, chapter, chapterIndex } = selectWeightedChapter(weightedChapters);
     const verse = chapter.Verses[Math.floor(Math.random() * chapter.Verses.length)];
 
     return {
